creditcoin: move types into separate module
